Move more duplicate methods to QueryFunctions
